Further refactor tidy up
