+++ lib/commands/import.js
@@ -22,6 +22,6 @@ const async = require('async');
 
   q.drain(async function() {
     console.log('Imported', Object.keys(forestLock).length, 'packages from IPFS')
-    await forest.closeDB()
+    await db.close()
   })
 })();
