+++ lib/commands/preload.js
@@ -68,7 +68,7 @@ const async = require('async');
 
       q.drain(async function() {
         console.log('Imported', packages.length, 'packages from', files.length, 'files')
-        await forest.closeDB()
+        await db.close()
       })
     })
