+++ lib/commands/republish.js
@@ -47,7 +47,7 @@ const async = require('async');
       fs.writeFileSync('forest.lock', JSON.stringify(cids, Object.keys(cids).sort(), 2), 'utf8')
       console.log(Object.keys(cids).length, 'Package CIDs written to forest.lock')
     }
-    await forest.closeDB()
+    await db.close()
   })
 
   fs.exists('package-lock.json', async (exists) => {
