+++ lib/commands/reset.js
@@ -4,8 +4,6 @@
 const forest = require('../forest');
 
 (async () => {
-  forest.connectDB()
   await forest.reset()
-  await forest.closeDB()
   console.log('Forest has been reset')
 })()
