+++ lib/commands/search.js
@@ -5,12 +5,12 @@ const forest = require('../forest');
 
 async function search(argv) {
   var query = argv.query
-  forest.connectDB()
+  var db = forest.connectDB()
   forest.search(query).then(async function(packages) {
     packages.forEach(function(pkg) {
       console.log(pkg.manager, pkg.name)
     });
-    await forest.closeDB()
+    await db.close()
   })
 }
