+++ lib/commands/sizes.js
@@ -5,11 +5,11 @@ const forest = require('../forest');
 
 var total = 0
 
-forest.connectDB()
-forest.filteredReadStream('size:').on('data', function (data) {
+var db = forest.connectDB()
+forest.core.filteredReadStream(db, 'size:').on('data', function (data) {
   console.log(data)
   total += parseInt(data.value)
 }).on('end', function () {
   console.log(total)
-  forest.closeDB()
+  db.close()
 })
