+++ lib/commands/update.js
@@ -22,7 +22,7 @@ const async = require('async');
 
     q.drain(async function() {
       console.log('Done')
-      await forest.closeDB()
+      await db.close()
     })
   })
 })();
