+++ lib/commands/verify.js
@@ -28,7 +28,7 @@ const async = require('async');
 
     q.drain(async function() {
       console.log('Verification complete!')
-      await forest.closeDB()
+      await db.close()
     })
   })
 })();
