+++ lib/forest.js
@@ -8,10 +8,6 @@ const toIterable = require('stream-to-it')
 const tmp = require('tmp-promise');
 const {zipDigest, zipManifest} = require('./zip_digest');
 
-const ssri = require('ssri')
-var crypto = require('crypto');
-const StreamZip = require('node-stream-zip');
-
 const core = require('./managers/core')
 const npm = require('./managers/npm')
 const go = require('./managers/go')
@@ -25,24 +21,16 @@ const packageAnnoucementsTopic = 'forest'
 
 var ipfsID = undefined;
 
-// db things
 function connectDB() {
   return db = level(paths.data)
 }
 
 async function reset() {
+  db = connectDB()
   await db.clear()
+  await db.close()
 }
 
-function filteredReadStream(start) {
-  return db.createReadStream({gte: start, lt: start+ '~'})
-}
-
-async function closeDB(){
-  return await db.close()
-}
-
-// ipfs things
 async function connectIPFS(db){
   try{
     ipfsID = await ipfs.id()
@@ -63,10 +51,6 @@ async function connectIPFS(db){
   }
 }
 
-
-
-
-// pubsub things
 async function defaultAnnounceCb(msg) {
   try{
     var json = JSON.parse(uint8ArrayToString(msg.data))
@@ -125,9 +109,6 @@ async function subscribePackageAnnoucements(receiveMsg = defaultAnnounceCb) {
   }
 }
 
-
-
-// commands
 async function seed(msg){
 
   if (ipfsID.id === msg.from) { return } // ignore our own announcements
@@ -149,8 +130,6 @@ async function importPackage(manager, name, version, resolved) {
   await core.loadMetadata(db, manager, name)
 }
 
-
-// core
 async function fetchAndAddtoIPFS(manager, name, version, url) {
   var hashAlg = 'sha2-256'
   var possibleCID
@@ -204,7 +183,7 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
 
          for (var i in lines) {
            // console.log('enqueing', lines[i])
-           var res = await addMetalineToIPFS(lines[i])
+           var res = await go.addMetalineToIPFS(lines[i])
            if(res){
              subfiles.push(res)
            }
@@ -287,7 +266,7 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
 
         metafile = await zipManifest(zippath)
         var metafileCID = await core.ipfsAdd(metafile, hashAlg)
-        var cids = await addZipContentsToIPFS(zippath)
+        var cids = await go.addZipContentsToIPFS(zippath)
 
         await fs.remove(zippath)
         cid = zipCID
@@ -304,73 +283,6 @@ async function fetchAndAddtoIPFS(manager, name, version, url) {
 
   return cid
 }
-
-// go specific
-async function addMetalineToIPFS(line){
-  try{
-
-    var parts = line.split('  ')
-    var f = {
-      sha256: parts[0],
-      path: parts[1],
-    }
-   console.log('Searching for', parts[1])
-   var digest = ssri.fromHex(parts[0], 'sha256', {single: true})
-   var maybeFileCID = core.guessCID(digest)
-
-   var individualFileContent = await core.attemptIPFSLoad(maybeFileCID)
-
-   if (individualFileContent){
-     f.cid = maybeFileCID
-     f.content = individualFileContent.toString()
-     console.log(f.path, 'guessed correctly')
-     return f
-   } else {
-     console.log('fail')
-     return false
-   }
- } catch(e) {
-   console.log('errrrrr', e)
-   return false;
- }
-}
-
-// go specific
-async function addZipContentsToIPFS(filePath) {
-  // TODO needs to be async
-  return new Promise((resolve, reject) => {
-  zip = new StreamZip({storeEntries: true, file: filePath});
-  zip.on('ready', async  () => {
-    cids = []
-    for (const entry of Object.values(zip.entries()).sort()) {
-        if(!entry.isDirectory){
-
-          const data = zip.entryDataSync(entry.name);
-          var cid = await core.ipfsAdd(data, 'sha2-256')
-          cids.push(cid)
-          console.log('Adding', entry.name, 'to ipfs', cid.cid.toString())
-        }
-    }
-    zip.close()
-    resolve(cids)
-  })
-  zip.on('error', e => {
-    try {
-      if (zip) {
-        zip.close()
-      }
-    } catch (e) {
-      // ignore
-    }
-    reject(e)
-  })
-});
-}
-
-
-
-
-
 
 function listPackages(manager) {
   if(manager){
@@ -414,8 +326,6 @@ function listPackageNames(manager) {
   })
 }
 
-
-
 async function downloadPackageFromIPFS(manager, name, version, cid) {
   try {
     existing_cid = await db.get(`cid:${manager}:${name}:${version}`)
@@ -436,7 +346,6 @@ async function downloadPackageFromIPFS(manager, name, version, cid) {
     console.log('Unknown manager:', manager)
   }
 }
-
 
 async function validate(manager, name, version) {
   try{
@@ -509,8 +418,6 @@ async function removePackage(manager, name, version) {
   await db.del(`cid:${manager}:${name}:${version}`)
 }
 
-// general things
-
 async function update(manager, name) {
   if (manager === 'npm'){
     var update = await core.updateMetadata('npm', name)
@@ -538,22 +445,17 @@ async function update(manager, name) {
   }
 }
 
-
-
 async function search(query) {
   // TODO allow filtering by manager
   var names = await listPackageNames('npm')
   return names.filter(function (pkg) { return pkg.name.indexOf(query) > -1; });
 }
 
-
-
 module.exports = {
   go,
   npm,
   core,
 
-  // public
   update,
   search,
   seed,
@@ -567,12 +469,8 @@ module.exports = {
   validate,
   importPackage,
   packageAnnoucementsTopic,
-
-  // not currently internal but should be
   downloadPackageFromIPFS,
   connectIPFS,
   connectDB,
-  closeDB,
-  filteredReadStream,
-  fetchAndAddtoIPFS // used inside republish command
+  fetchAndAddtoIPFS
 }
