+++ lib/managers/core.js
@@ -288,6 +288,10 @@ async function dialPeers(db) {
   })
 }
 
+function filteredReadStream(db, start) {
+  return db.createReadStream({gte: start, lt: start+ '~'})
+}
+
 module.exports = {
   loadMetadata,
   saveCid,
@@ -302,5 +306,6 @@ module.exports = {
   exportPackages,
   listPeers,
   dialPeers,
-  loadVersionMetadata
+  loadVersionMetadata,
+  filteredReadStream
 }
