+++ lib/managers/go.js
@@ -2,12 +2,16 @@ const tmp = require('tmp-promise');
 const fetch = require('node-fetch')
 const fs = require('fs-extra');
 const path = require('path')
+
 const pipe = require('it-pipe')
 const toIterable = require('stream-to-it')
 const IpfsHttpClient = require('ipfs-http-client')
 const { urlSource } = IpfsHttpClient
 const ipfs = IpfsHttpClient()
 
+const ssri = require('ssri')
+
+const StreamZip = require('node-stream-zip');
 const {zipDigest, zipManifest} = require('../zip_digest');
 
 const core = require('./core')
@@ -396,6 +400,64 @@ async function watchImporter(db, change) {
   }
 }
 
+async function addMetalineToIPFS(line){
+  try{
+    var parts = line.split('  ')
+    var f = {
+      sha256: parts[0],
+      path: parts[1],
+    }
+   console.log('Searching for', parts[1])
+   var digest = ssri.fromHex(parts[0], 'sha256', {single: true})
+   var maybeFileCID = core.guessCID(digest)
+
+   var individualFileContent = await core.attemptIPFSLoad(maybeFileCID)
+
+   if (individualFileContent){
+     f.cid = maybeFileCID
+     f.content = individualFileContent.toString()
+     console.log(f.path, 'guessed correctly')
+     return f
+   } else {
+     console.log('fail')
+     return false
+   }
+ } catch(e) {
+   console.log('errrrrr', e)
+   return false;
+ }
+}
+
+async function addZipContentsToIPFS(filePath) {
+  // TODO needs to be async
+  return new Promise((resolve, reject) => {
+  zip = new StreamZip({storeEntries: true, file: filePath});
+  zip.on('ready', async  () => {
+    cids = []
+    for (const entry of Object.values(zip.entries()).sort()) {
+        if(!entry.isDirectory){
+
+          const data = zip.entryDataSync(entry.name);
+          var cid = await core.ipfsAdd(data, 'sha2-256')
+          cids.push(cid)
+          console.log('Adding', entry.name, 'to ipfs', cid.cid.toString())
+        }
+    }
+    zip.close()
+    resolve(cids)
+  })
+  zip.on('error', e => {
+    try {
+      if (zip) {
+        zip.close()
+      }
+    } catch (e) {
+      // ignore
+    }
+    reject(e)
+    })
+  });
+}
 
 module.exports = {
   parseName,
@@ -417,5 +479,7 @@ module.exports = {
   fetchLatest,
   downloadPackageFromRegistry,
   watchImporter,
-  serverHandler
+  serverHandler,
+  addMetalineToIPFS,
+  addZipContentsToIPFS
 }
